pyproject.toml
[project]
name = "axf-servo-daq"
version = "1.0.0"
description = "The servo test bench DAQ software"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyinstaller>=6.20.0",
    "pyqt6>=6.11.0",
    "pyqtgraph>=0.14.0",
    "pyserial>=3.5",
    "pyside6>=6.11.0",
    "qt6-tools>=6.5.0.1.3",
]
